Create install_dependencies.sh
This script updates the package manager, installs Python and its dependencies, installs Node.js and npm, installs TensorFlow and other Python packages, installs Airflow and its dependencies, installs Node-RED, and creates the airflow home directory.

The script installs the TensorFlow package with version 2.4.0, but you can update this to the latest version or a specific version that you want to use.

The script installs the postgres and redis extras for Airflow, which are required for the webserver service in the docker-compose.yml file.

The script installs Node-RED globally using the -g flag and sets the --unsafe-perm flag to allow the installation to run with root privileges. This is necessary because the script is being run as the root user.

Finally, the script creates the /usr/local/airflow directory and sets the correct permissions to allow the airflow user to access it. This is necessary because the webserver service in the docker-compose.yml file mounts the ./scripts directory as a volume in the /usr/local/airflow directory.