Create download_model.sh
This script uses curl to download the model from a URL and save it as model.tar.gz. It then uses tar to extract the contents of model.tar.gz to the ./models/tensorflow directory. Finally, it removes the tarball to save space.

You will need to replace the URL with the URL of the model that you want to download and the ./models/tensorflow directory with the directory where you want to save the model.

You may also want to add error handling to the script to handle cases where the download fails or the tarball is corrupt.