Create export_model.py
The export_model function takes two arguments: a model and an export_path string specifying the path where the model should be exported. It exports the model as a TensorFlow SavedModel using the save method from the TensorFlow saved_model module.

The if __name__ == '__main__': block loads the model weights using the load_weights method and creates the model using the create_model function. It then calls the export_model function to export the model as a TensorFlow SavedModel.

The script assumes that the model weights have been saved in the ./models/tensorflow directory, and that the input shape of the model is (224, 224, 3), which is the size of the input images.

The script exports the model as a TensorFlow SavedModel in the ./models/tensorflow/saved_model directory. You can modify the directory where the weights are saved and the directory where the model should be exported as needed.