Create model_training_pipeline.py
The script defines a Directed Acyclic Graph (DAG) using the Airflow DAG class and five tasks using the Airflow PythonOperator class. Each task is defined as a Python function that performs a specific step in the model training pipeline.

The default_args dictionary contains default arguments for the DAG, such as the owner, start date, and retry settings. The dag object is created using the DAG class and the default arguments. The schedule_interval argument specifies how often the DAG should be run.

The tasks are defined as Python functions that perform a specific step in the model training pipeline. For example, the download_data function downloads the data, the preprocess_data function preprocesses the data, and so on. Each task is created using the PythonOperator class and the dag object.

The dependencies between the tasks are set using the >> operator. This specifies the order in which the tasks should be run. In this example, the tasks are run in the following order: task_1 (download data) -> task_2 (preprocess data) -> task_3 (train model) -> task_4 (evaluate model) -> task_5 (export model).

This script can be saved in the /airflow/dags/ directory and run using the Airflow scheduler. The scheduler will execute the tasks in the specified order according to the schedule interval specified in the dag object. You can modify the tasks and dependencies as needed to fit your specific model training pipeline.